# HTTP server connection keeps processing pending pipelined requests