Fix for #2365