@vietj 2018-04-04
now I'm wondering about the current code:

```
  synchronized void handleMessage(Object msg) {
    if (queueing) {
      enqueue(msg);
    } else {
      if (processMessage(msg)) {
        checkNextTick();
      } else {
        enqueue(msg);
      }
    }
  }
```

that seems correct, but I think that it's not necessary to call `checkNextTick()` if the request process the message, i.e:

```
  synchronized void handleMessage(Object msg) {
    if (queueing) {
      enqueue(msg);
    } else {
      if (processMessage(msg)) {
        // Since we processed the message it means that the queue is empty so we don't need to check the next tick
        checkNextTick();
      } else {
        enqueue(msg);
      }
    }
  }
```