@vietj 2018-04-04
so we could have instead

```
  synchronized void handleMessage(Object msg) {
    if (queueing || !processMessage(msg)) {
      enqueue(msg);
    }
  }
```