@vietj 2018-04-04
@tsegismont I've updated the PR to not call `checkNextTick()` when a message is processed in `handleMessage(Object)` and improved the related tests