@vietj 2018-04-05
the microbenchmark for HTTP server shows significant improvements likely related to the removal of the unecessary `checkNextTick` invocation in `handleMessage`:

### this branch

```
HttpServerHandlerBenchmark.netty     thrpt   10  840.095 ± 132.653  ops/ms
HttpServerHandlerBenchmark.vertx     thrpt   10  649.268 ±  10.494  ops/ms
HttpServerHandlerBenchmark.vertxOpt  thrpt   10  711.100 ±   9.458  ops/ms
```

### master branch

```
HttpServerHandlerBenchmark.netty     thrpt   10  845.225 ± 85.887  ops/ms
HttpServerHandlerBenchmark.vertx     thrpt   10  620.237 ±  9.065  ops/ms
HttpServerHandlerBenchmark.vertxOpt  thrpt   10  690.147 ± 11.093  ops/ms
```