@tsegismont 2018-04-05
>@tsegismont the server connections receives pipelined requests messages (HttpRequest, HttpLastContent, HttpRequest, etc...)
>
>The first requests does not end the first response so the field pendingResponse remains assigned and it closes the connection.
>
>When the second requests arrives (it arrives in the same channel read operation since it's pipelined and aggregated at the TCP level) then processMessage(Object) returns false and the request object is enqueued. The first processing message triggered a call to checkNextTick(), when the asynchronous checkNextTick() runs it tries to process the message and fails since pendingResponse is still assigned, the message is put back on the queue and since the queue is not empty checkNextTick() is called again and so on.

I understand that but I can't see why the new test would fail without the fix.