@vietj 2018-04-05
ok, so what will fail is not the test itself but the shutdown of the corresponding Vert.x instance because the event loop will continue to process tasks and the vertx.close(Handler<AsyncResult>> will not complete in time.

I can add a note in the test about it.