@tsegismont 2018-04-05
>I can add a note in the test about it.

:+1: